+++ CHANGES.rst
@@ -1,1 +1,4 @@
 - Use Select2 widget for the document type selection form.
+- Update source column matching to be additive and not exclusive.
+- Add two columns to show the number of documents per workflow and
+  workflow state.